Refactor ChatManager
